토큰을 사용하지 않았을 때 401상태를 보내도록 변경

+++ server/src/main/java/wap/starlist/bookmark/domain/Bookmark.java
@@ -43,7 +43,7 @@ public class Bookmark {
 
     private Boolean syncing;
 
-    private Integer googleId;
+    private Long googleId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "folder_id")
